+++ Game1.cs
@@ -53,10 +53,11 @@ namespace GordonWare
         protected override void Initialize()
         {
             // If you want the game to only play your minigame in order to test it, please remove every other game and add only yours to MiniGameManager
-            // MiniGameManager.AddMiniGame(TemplateGame));
+            //MiniGameManager.AddMiniGame(TemplateGame));
             MiniGameManager.AddMiniGame(new KeyboardGame());
             MiniGameManager.AddMiniGame(new DabGame());
             MiniGameManager.AddMiniGame(new TaupeGame());
+            MiniGameManager.AddMiniGame(new ChooseService());
 
             base.Initialize();
         }
