+++ GordonWare.csproj
@@ -43,6 +43,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="Calendar.cs" />
+    <Compile Include="ChooseService.cs" />
     <Compile Include="Game1.cs" />
     <Compile Include="Input.cs" />
     <Compile Include="DabGame.cs" />
